airflow/requirements.txt
apache-airflow-providers-google
google-auth
pyarrow
python-dotenv
pandas==1.5.2
protobuf==4.21.11
pandas-gbq==0.18.1
sqlalchemy==1.4.46


